Fix QbeastStats serialization/deserialization tests
